requirements.txt
asgiref==3.7.2
channels==3.0.5
Django==4.2.6
Pillow==10.1.0
Requests==2.31.0
django-cors-headers==4.3.0